Finished hermes protocol without configured generating plugin
