+++ build.gradle
@@ -1,4 +1,6 @@
-apply plugin: 'idea'
+plugins {
+    id 'idea'
+}
 
 allprojects {
     group 'org.hermes'
@@ -26,6 +28,7 @@ subprojects {
         testCompile group: 'junit', name: 'junit', version: '4.12'
         implementation group: 'org.apache.avro', name: 'avro', version: '1.10.1'
         implementation group: 'io.confluent', name: 'kafka-avro-serializer', version: '6.1.0'
+        implementation group: 'org.slf4j', name: 'slf4j-simple', version: '1.7.30'
     }
 
     checkstyle {
