+++ develop-producer/src/avro/hermes.avsc
@@ -1,0 +1,50 @@
+{
+  "type": "record",
+  "namespace": "org.hermes.developProducer.avro",
+  "name": "HermesRecord",
+  "version": "1",
+  "fields": [
+    {
+      "name": "data",
+      "type": "string",
+      "doc": "The data which will be scraping from",
+      "default": ""
+    },
+    {
+      "name": "fields",
+      "type": {
+        "type": "array",
+        "items": {
+          "name": "Field",
+          "type": "record",
+          "fields": [
+            {
+              "name": "selector",
+              "type": "string",
+              "doc": "Selector is the rule showing how the core will try to get the desired field"
+            },
+            {
+              "name": "output_type",
+              "type": {
+                "type": "enum",
+                "name": "Enum_type",
+                "symbols": [
+                  "SINGLE",
+                  "MULTIPLE"
+                ]
+              },
+              "default": "MULTIPLE",
+              "doc": "Enum defines object will be single object or array"
+            },
+            {
+              "name": "output_name",
+              "type": "string",
+              "doc": "Name of the output field"
+            }
+          ]
+        }
+      },
+      "doc": "The array of the fields, that describe the output of analysed data"
+    }
+  ]
+}
