--- develop-producer/src/main/java/org/hermes/developProducer/DevelopProducerApplication.java
@@ -1,11 +1,0 @@
-package org.hermes.developProducer;
-
-import org.hermes.developProducer.service.KafkaProducerService;
-
-public class DevelopProducerApplication {
-
-    public static void main(String[] args) {
-        var kafkaProducerService = new KafkaProducerService();
-        kafkaProducerService.runProducer();
-    }
-}
