--- develop-producer/src/main/java/org/hermes/developProducer/config/KafkaConfig.java
@@ -1,9 +1,0 @@
-package org.hermes.developProducer.config;
-
-public class KafkaConfig {
-
-    public static final String TOPIC = "develop-producer";
-    public static final String KAFKA_SERVER_URL = "localhost";
-    public static final int KAFKA_SERVER_PORT = 9092;
-
-}
