+++ develop-producer/src/main/java/org/hermes/developproducer/DevelopProducerApplication.java
@@ -1,0 +1,11 @@
+package org.hermes.developproducer;
+
+import org.hermes.developproducer.service.KafkaProducerService;
+
+public class DevelopProducerApplication {
+
+    public static void main(String[] args) {
+        KafkaProducerService kafkaProducerService = new KafkaProducerService();
+        kafkaProducerService.runProducer();
+    }
+}
