+++ develop-producer/src/main/java/org/hermes/developproducer/config/KafkaConfig.java
@@ -1,0 +1,14 @@
+package org.hermes.developproducer.config;
+
+public class KafkaConfig {
+
+    public static final String TOPIC = "develop-producer";
+
+    public static final String KAFKA_SERVER_URL = "127.0.0.1";
+    public static final int KAFKA_SERVER_PORT = 9092;
+
+    public static final String PRODUCER_KEY = "develop-producer";
+
+    public static final String SCHEMA_REGISTRY_URL = "http://localhost";
+    public static final int SCHEMA_REGISTRY_PORT = 8081;
+}
