+++ develop-producer/src/test/java/org/hermes/developproducer/DevelopProducerApplicationTest.java
@@ -1,4 +1,4 @@
-package org.hermes.developProducer;
+package org.hermes.developproducer;
 
 import org.junit.Test;
